# Conflicts:
#	src/main/groovy/com/netgrif/workflow/startup/ElasticsearchRunner.groovy
#	src/main/java/com/netgrif/workflow/elastic/domain/ElasticCase.java

+++ src/main/java/com/netgrif/workflow/workflow/web/WorkflowController.java
@@ -1,6 +1,10 @@
 package com.netgrif.workflow.workflow.web;
 
+import com.netgrif.workflow.auth.domain.Author;
 import com.netgrif.workflow.auth.domain.LoggedUser;
+import com.netgrif.workflow.auth.domain.User;
+import com.netgrif.workflow.auth.domain.throwable.UnauthorisedRequestException;
+import com.netgrif.workflow.auth.service.interfaces.IUserService;
 import com.netgrif.workflow.elastic.domain.ElasticCase;
 import com.netgrif.workflow.elastic.service.interfaces.IElasticCaseService;
 import com.netgrif.workflow.elastic.web.CaseSearchRequest;
@@ -54,6 +58,9 @@ public class WorkflowController {
     @Autowired
     private IDataService dataService;
 
+    @Autowired
+    private IUserService userService;
+
     @RequestMapping(value = "/case", method = RequestMethod.POST)
     public CaseResource createCase(@RequestBody CreateCaseBody body, Authentication auth) {
         LoggedUser loggedUser = (LoggedUser) auth.getPrincipal();
@@ -106,7 +113,12 @@ public class WorkflowController {
     }
 
     @RequestMapping(value = "/case/{id}", method = RequestMethod.DELETE)
-    public MessageResource deleteCase(@PathVariable("id") String caseId) {
+    public MessageResource deleteCase(@PathVariable("id") String caseId) throws UnauthorisedRequestException {
+        User logged = userService.getLoggedUser();
+        Case requestedCase = workflowService.findOne(caseId);
+        if( !logged.transformToLoggedUser().isAdmin() && !logged.getId().equals(requestedCase.getAuthor().getId()))
+            throw new UnauthorisedRequestException("User " + logged.transformToLoggedUser().getUsername() + " doesn't have permission to delete case " + requestedCase.getStringId());
+
         try {
             caseId = URLDecoder.decode(caseId, StandardCharsets.UTF_8.name());
             workflowService.deleteCase(caseId);
